+++ 0x01-warm_up/search_algos.h
@@ -1,7 +1,9 @@
 #ifndef _SEARCH_ALGOS_H_
 #define _SEARCH_ALGOS_H_
 
+#include <stdio.h>
+
 int binary_search(int *array, size_t size, int value);
-void printer(int *array, int i, int size);
+void printer(int *array, size_t i, size_t size);
 
 #endif
